requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.84
botocore==1.34.84
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-rest-knox==4.2.0
django-sslserver==0.22
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jax==0.4.26
jmespath==1.0.1
keras==2.12.0
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.4.0
msgpack==1.0.8
numpy==1.23.5
oauthlib==3.2.2
openai==1.17.1
opt-einsum==3.3.0
packaging==24.0
priority==1.3.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
scipy==1.13.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tqdm==4.66.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
zope.interface==6.3
